+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.model.Film;
-import ru.yandex.practicum.filmorate.service.FilmService;
+import ru.yandex.practicum.filmorate.storage.InMemoryFilmStorage;
 
 import javax.validation.Valid;
 import java.util.List;
@@ -14,22 +14,22 @@ import java.util.List;
 @AllArgsConstructor
 @RequestMapping("/films")
 public class FilmController {
-    private final FilmService filmService;
+    private final InMemoryFilmStorage inMemoryFilmStorage;
 
     @GetMapping
     public List<Film> getAllFilms() {
-        return filmService.getAllFilms();
+        return inMemoryFilmStorage.getAllFilms();
     }
 
     @PostMapping
     public Film addFilm(@Valid @RequestBody Film film) {
         log.info("Идет добавление фильма: " + film.getName());
-        return filmService.addFilm(film);
+        return inMemoryFilmStorage.addFilm(film);
     }
 
     @PutMapping
     public Film updateFilm(@Valid @RequestBody Film film) {
         log.info("Идет обновление фильма: " + film.getName());
-        return filmService.updateFilm(film);
+        return inMemoryFilmStorage.updateFilm(film);
     }
 }
