+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -6,6 +6,7 @@ import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Past;
 import java.time.LocalDate;
+import java.util.Set;
 
 @Data
 public class User {
@@ -17,4 +18,5 @@ public class User {
     private String name;
     @Past
     private LocalDate birthday;
+    private Set<Long> friendsId;
 }
