+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -1,49 +1,24 @@
 package ru.yandex.practicum.filmorate.service;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
+import ru.yandex.practicum.filmorate.storage.UserStorage;
 
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 @Service
 public class UserService {
-    private int id = 1;
-    private final Map<Integer, User> users = new HashMap<>();
 
-    public List<User> getAllUsers() {
-        List<User> allUsers = new ArrayList<>();
-        for (Map.Entry<Integer, User> user : users.entrySet()) {
-            allUsers.add(user.getValue());
-        }
-        return allUsers;
+    private final UserStorage userStorage;
+
+    @Autowired
+    public UserService(UserStorage userStorage) {
+        this.userStorage = userStorage;
     }
 
-    private int setId() {
-        return id++;
-    }
-
-    public User createUser(User user) {
-        if (users.containsKey(user.getId())) {
-            throw new ValidationException("Пользователь с логином " + user.getLogin() +
-                    " уже зарегистрирован.");
-        }
-        user.setId(setId());
-        if (user.getName() == null || user.getName().isBlank()) {
-            user.setName(user.getLogin());
-        }
-        users.put(user.getId(), user);
-        return user;
-    }
-
-    public User updateUser(User user) {
-        if (!users.containsKey(user.getId())) {
-            throw new ValidationException("Пользователя с id " + user.getId() + " не существует.");
-        }
-        users.put(user.getId(), user);
-        return user;
+    public Set<Long> getFriends(Integer id) {
+        return userStorage.getAllUsers().get(id).getFriendsId();
     }
 }
